Создает модуль описания фотографии и модуль вспомогательных функций
